+++ Benchmarks/src/Collections/List/ImmutableListBenchmarks.cs
@@ -7,8 +7,9 @@ namespace Benchmarks.Collections.List;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class ImmutableListBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	public static readonly ImmutableList<int> Data;
 
@@ -19,7 +20,7 @@ public class ImmutableListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values sequentially from a ImmutableList")]
 	public static int ImmutableListGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data[CollectionsHelpers.SequentialIndices[j]];
 			}
@@ -31,7 +32,7 @@ public class ImmutableListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values randomly from a ImmutableList")]
 	public static int ImmutableListGetRandom() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data[CollectionsHelpers.RandomIndices[j]];
 			}
@@ -46,7 +47,7 @@ public class ImmutableListBenchmarks {
 		int result = 0;
 		ImmutableList<int> target = ImmutableList<int>.Empty;
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int index = 0; index < Data.Count; index++) {
 				target = target.Add(index);
 			}
@@ -63,7 +64,7 @@ public class ImmutableListBenchmarks {
 		int result = 0;
 		ImmutableList<int> target = ImmutableList<int>.Empty;
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				target = target.Add(j);
 			}
@@ -81,7 +82,7 @@ public class ImmutableListBenchmarks {
 	[Benchmark("ListCreation", "Tests allocation and initialization of an ImmutableList")]
 	public static int ImmutableListCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableList<int> immutableList = ImmutableList<int>.Empty;
 			for (int index = 0; index < Data.Count; index++) {
 				immutableList = immutableList.Add(index * 2);
@@ -97,7 +98,7 @@ public class ImmutableListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an ImmutableList using a foreach loop")]
 	public static int ImmutableListCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableList<int> target = ImmutableList<int>.Empty;
 			foreach (int element in Data) {
 				target = target.Add(element);
@@ -113,7 +114,7 @@ public class ImmutableListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an ImmutableList using a for loop")]
 	public static int ImmutableListCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableList<int> target = ImmutableList<int>.Empty;
 			for (int index = 0; index < Data.Count; index++) {
 				target = target.Add(Data[index]);
