+++ Benchmarks/src/Collections/Table/HashtableBenchmarks.cs
@@ -9,8 +9,9 @@ namespace Benchmarks.Collections.Table;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class HashtableBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	public static readonly Hashtable Data = new(1000);
 
@@ -24,7 +25,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableInsertion", "Tests insertion into a Hashtable")]
 	public static int HashtableInsertion() {
 		Hashtable temp = new Hashtable();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			temp = new Hashtable();
 			for (int j = 0; j < 1000; j++) {
 				temp.Add(CollectionsHelpers.SequentialIndices[j], CollectionsHelpers.RandomValues[j]);
@@ -37,7 +38,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableGet", "Tests getting values sequentially from a Hashtable")]
 	public static int HashtableGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += (int)Data[CollectionsHelpers.SequentialIndices[j]];
 			}
@@ -49,7 +50,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableGet", "Tests getting values randomly from a Hashtable")]
 	public static int HashtableGetRandom() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += (int)Data[CollectionsHelpers.RandomIndices[j]];
 			}
@@ -61,7 +62,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableRemoval", "Tests removal from a Hashtable")]
 	public static int HashtableRemoval() {
 		Hashtable temp = new Hashtable();
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < 1000; j++) {
 				temp.Add(CollectionsHelpers.SequentialIndices[j], CollectionsHelpers.RandomValues[j]);
 			}
@@ -77,7 +78,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableCreation", "Tests allocation and initialization of a Hashtable")]
 	public static int HashtableCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Hashtable hashtable = new Hashtable();
 			for (int index = 0; index < Data.Count; index++) {
 				hashtable.Add(index, index * 2);
@@ -92,7 +93,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableCopy", "Tests copying a Hashtable using a foreach loop looping through k/v pairs")]
 	public static int HashtableCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Hashtable target = new Hashtable();
 
 			foreach (DictionaryEntry pair in Data) {
@@ -109,7 +110,7 @@ public class HashtableBenchmarks {
 	[Benchmark("TableCopy", "Tests copying a Hashtable using a foreach loop and looping through keys")]
 	public static int HashtableCopyManualForeachIndex() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			Hashtable target = new Hashtable();
 
 			foreach (int key in Data.Keys) {
