+++ Benchmarks/src/HelperObjects/Inheritance/IDoable.cs
@@ -2,5 +2,5 @@
 
 public interface IDoable {
 
-	int UpdateAndGetValue();
+	ulong UpdateAndGetValue();
 }
